refactor(fgo): extract FGOConfig (1.5k lines) from fgo.hpp into fgo_config.hpp
The config struct alone was 57% of the 2.7k-line header. It now lives
as fgo::Config in <libgnss++/algorithms/fgo_config.hpp>; FGOBackend
moved alongside it since it parameterizes the backend knob. The class
keeps the historical FGOProcessor::FGOConfig name via an alias, so all
171 qualified uses compile unchanged.
